+++ src/Javis.jl
@@ -67,6 +67,8 @@ Create a video with a certain `width` and `height` in pixel.
 This also sets `CURRENT_VIDEO`.
 """
 function Video(width, height)
+    # some luxor functions need a drawing ;)
+    Drawing()
     video = Video(width, height, Dict{Symbol,Any}())
     if isempty(CURRENT_VIDEO)
         push!(CURRENT_VIDEO, video)
@@ -164,6 +166,7 @@ A SubAction should not be created by hand but instead by using one of the constr
 - `transitions::Vector{Transition}`: A list of transitions like [`Translation`](@ref)
 - `internal_transitions::Vector{InternalTransition}`:
     A list of internal transitions which store the current transition for a specific frame.
+- `defs::Dict{Symbol, Any}` any kind of definitions that are relevant for the subaction.
 """
 mutable struct SubAction <: AbstractAction
     frames::Frames
@@ -171,6 +174,7 @@ mutable struct SubAction <: AbstractAction
     func::Function
     transitions::Vector{Transition}
     internal_transitions::Vector{InternalTransition}
+    defs::Dict{Symbol,Any}
 end
 
 SubAction(transitions::Transition...) = SubAction(:same, transitions...)
@@ -281,7 +285,7 @@ SubAction(frames, trans::Transition...) =
 
 
 SubAction(frames, anim::Animation, func::Function, transitions::Transition...) =
-    SubAction(frames, anim, func::Function, collect(transitions), [])
+    SubAction(frames, anim, func::Function, collect(transitions), [], Dict{Symbol,Any}())
 
 """
     ActionSetting
@@ -1428,7 +1432,7 @@ export Video, Action, BackgroundAction, SubAction, Rel
 export Line, Translation, Rotation, Transformation, Scaling
 export val, pos, ang, get_value, get_position, get_angle
 export projection, morph
-export appear, disappear, rotate_around
+export appear, disappear, rotate_around, follow_path
 export rev
 export scaleto